+++ database/migrations/2025_04_21_061552_project_locations.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create("project_locations", function (Blueprint $table) {
             $table->id();
-            $table->integer("project_id");
+            $table->unsignedBigInteger("project_id");
             $table->string("province");
             $table->string("city");
             $table->string("address");
